+++ src/com/thegongoliers/output/EnhancedRobotDrive.java
@@ -1,5 +1,7 @@
 package com.thegongoliers.output;
 
+import com.thegongoliers.output.interfaces.MecanumDriveTrainInterface;
+
 import edu.wpi.first.wpilibj.RobotDrive;
 import edu.wpi.first.wpilibj.SpeedController;
 import edu.wpi.first.wpilibj.interfaces.Gyro;
